Move form-associated classes to separate files
