+++ packages/web-components/fast-foundation/src/radio/radio.form-associated.ts
@@ -1,0 +1,18 @@
+import { FASTElement } from "@microsoft/fast-element";
+import { FormAssociated } from "../form-associated/form-associated";
+
+/**
+ * A form-associated base class for the {@link (Radio:class)} component.
+ *
+ * @internal
+ */
+export class FormAssociatedRadio extends FormAssociated(
+    class extends FASTElement {
+        public proxy: HTMLInputElement = document.createElement("input");
+    }
+) {}
+
+/**
+ * @internal
+ */
+export interface FormAssociatedRadio extends FormAssociated {}
