+++ packages/web-components/fast-foundation/src/slider/slider.form-associated.ts
@@ -1,0 +1,18 @@
+import { FASTElement } from "@microsoft/fast-element";
+import { FormAssociated } from "../form-associated/form-associated";
+
+/**
+ * A form-associated base class for the {@link (Slider:class)} component.
+ *
+ * @internal
+ */
+export class FormAssociatedSlider extends FormAssociated(
+    class extends FASTElement {
+        public proxy: HTMLInputElement = document.createElement("input");
+    }
+) {}
+
+/**
+ * @internal
+ */
+export interface FormAssociatedSlider extends FormAssociated {}
