+++ packages/web-components/fast-foundation/src/switch/switch.form-associated.ts
@@ -1,0 +1,18 @@
+import { FASTElement } from "@microsoft/fast-element";
+import { FormAssociated } from "../form-associated/form-associated";
+
+/**
+ * A form-associated base class for the {@link (Switch:class)} component.
+ *
+ * @internal
+ */
+export class FormAssociatedSwitch extends FormAssociated(
+    class extends FASTElement {
+        public proxy: HTMLInputElement = document.createElement("input");
+    }
+) {}
+
+/**
+ * @internal
+ */
+export interface FormAssociatedSwitch extends FormAssociated {}
